submission/Deployment/object-detection-database/requirements.txt
absl-py==2.0.0
acme==1.21.0
ann-visualizer==2.5
anyio==4.0.0
anylabeling==0.2.23
apturl==0.5.2
asttokens==2.2.1
astunparse==1.6.3
backcall==0.2.0
bardapi==0.1.38
bcrypt==3.2.0
beautifulsoup4==4.12.2
blinker==1.4
Brlapi==0.8.3
browser-cookie3==0.19.1
cachetools==5.0.0
certbot==1.21.0
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
colorama==0.4.4
coloredlogs==15.0.1
comm==0.1.3
command-not-found==0.3
ConfigArgParse==1.5.3
configobj==5.0.6
contourpy==1.0.7
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
debugpy==1.6.7
decorator==5.1.1
deep-translator==1.11.4
defer==1.0.6
distlib==0.3.7
distro==1.7.0
distro-info==1.1+ubuntu0.1
duplicity==0.8.21
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==1.2.0
fasteners==0.14.1
filelock==3.12.2
flatbuffers==23.3.3
fonttools==4.39.3
fsspec==2023.10.0
future==0.18.3
gast==0.4.0
gitdb==4.0.11
GitPython==3.1.40
google-api-core==2.12.0
google-api-python-client==1.7.11
google-auth==2.23.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-cloud-core==2.3.3
google-cloud-translate==3.12.0
google-pasta==0.2.0
googleapis-common-protos==1.61.0
graphviz==0.20.1
greenlet==1.1.2
grpcio==1.59.0
grpcio-status==1.59.0
h11==0.14.0
h2==4.1.0
h5py==3.9.0
hpack==4.0.0
httpcore==0.18.0
httplib2==0.20.2
httpx==0.25.0
humanfriendly==10.0
hyperframe==6.0.1
hypothesis==6.82.4
idna==3.3
imbalanced-learn==0.11.0
imblearn==0.0
imgviz==1.7.2
importlib-metadata==4.6.4
ipykernel==6.23.0
ipython==8.13.2
jedi==0.18.2
jeepney==0.7.1
Jinja2==3.1.2
joblib==1.3.1
josepy==1.10.0
jupyter_client==8.2.0
jupyter_core==5.3.0
kazam==1.4.5
keras==2.13.1
keras-visualizer==3.1.2
keyring==23.5.0
kiwisolver==1.4.4
langdetect==1.0.9
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==16.0.6
lockfile==0.12.2
louis==3.20.0
lxml==4.9.2
lz4==4.3.2
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
natsort==8.3.1
nest-asyncio==1.5.6
netifaces==0.11.0
networkx==3.2.1
np-utils==0.6.0
numpy==1.24.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.52
nvidia-nvtx-cu12==12.1.105
oauth2client==4.1.3
oauthlib==3.2.0
olefile==0.46
onnx==1.13.1
onnxruntime==1.14.1
opencv-python==4.8.1.78
opencv-python-headless==4.7.0.72
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.3
paramiko==2.9.3
parsedatetime==2.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.5.1
pptx2md==1.3.0
prompt-toolkit==3.0.38
proto-plus==1.22.3
protobuf==4.24.4
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycairo==1.20.1
pycryptodomex==3.11.0
pycups==2.0.1
pydot==1.4.2
PyDrive==1.3.1
Pygments==2.15.1
PyGObject==3.42.1
PyICU==2.8.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
pyRFC3339==1.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-pptx==0.6.21
pytz==2022.1
pyxattr==0.7.2
pyxdg==0.27
PyYAML==5.4.1
pyzmq==25.0.2
qimage2ndarray==1.10.0
rapidfuzz==3.0.0
regex==2023.8.8
reportlab==3.6.8
requests==2.25.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rsa==4.8
scikit-learn==1.3.1
scikit-plot==0.3.7
scipy==1.11.2
screen-resolution-extra==0.0.0
seaborn==0.12.2
SecretStorage==3.3.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==1.4.31
stack-data==0.6.2
sympy==1.11.1
systemd-python==234
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
torch==2.1.0
torch-utils==0.1.2
torchvision==0.16.0
tornado==6.3.1
tqdm==4.65.0
traitlets==5.9.0
triton==2.1.0
typing==3.7.4.3
typing_extensions==4.5.0
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
ultralytics==8.0.207
unattended-upgrades==0.1
uritemplate==3.0.1
urllib3==1.26.5
usb-creator==0.3.7
virtualenv==20.24.3
wadllib==1.3.6
wcwidth==0.2.6
Werkzeug==2.3.7
wrapt==1.15.0
xdg==5
xkit==0.0.0
XlsxWriter==3.1.0
youtube-dl==2021.12.17
zipp==1.0.0
zope.component==4.3.0
zope.event==4.4
zope.hookable==5.1.0
zope.interface==5.4.0
